+++ packages/ui/src/components/organization/personal-account-view.tsx
@@ -87,7 +87,7 @@ export function PersonalAccountView({
                     <>
                         <span
                             className={cn(
-                                "truncate font-semibold",
+                                "truncate font-medium",
                                 size === "lg" ? "text-base" : "text-sm",
                                 classNames?.title
                             )}
@@ -105,8 +105,8 @@ export function PersonalAccountView({
                         {size !== "sm" && (
                             <span
                                 className={cn(
-                                    "truncate opacity-70",
-                                    size === "lg" ? "text-sm" : "text-xs",
+                                    "truncate text-xs text-muted-foreground",
+                                    size === "lg" ? "text-sm" : "",
                                     classNames?.subtitle
                                 )}
                             >