+++ packages/ui/src/components/ui/user-view.tsx
@@ -108,7 +108,7 @@ export function UserView({
                     <>
                         <span
                             className={cn(
-                                "truncate font-semibold",
+                                "truncate font-medium",
                                 size === "lg" ? "text-base" : "text-sm",
                                 classNames?.title
                             )}
@@ -128,8 +128,8 @@ export function UserView({
                             (user?.name || user?.username) && (
                                 <span
                                     className={cn(
-                                        "truncate opacity-70",
-                                        size === "lg" ? "text-sm" : "text-xs",
+                                        "truncate text-xs text-muted-foreground",
+                                        size === "lg" ? "text-sm" : "",
                                         classNames?.subtitle
                                     )}
                                 >